feat(sampler): coordinate-decoupled rollback, 32-bit seed ABI, RNG continuity & state invalidation (v1.6.3)
- Coordinate-Decoupled Rollback: Introduce struct argus_history_entry to tag history tokens with evaluated KV positions. Primed prompt tokens (kv_pos = -1) survive KV rollbacks while orphaned branch tokens (kv_pos >= start_pos && kv_pos >= 0) are pruned.
- 32-Bit Seed ABI & Explicit Padding: Align argus_sampler_params_t with upstream 32-bit llama_sampler_init_dist using uint32_t seed and uint8_t reserved_padding[4], preserving the exact 56-byte 8-byte-aligned layout across C and Java Panama FFM.
- Persistent RNG Stream Continuity: Preserve and transfer active llama_sampler_dist instances across non-seed hyperparameter and logit bias reconfigurations without restarting the RNG stream from draw #0.
- Unified Stale Logits Invalidation: Enforce invalidate_seq_logits() across KV cache mutations, decode errors, multimodal errors, and sampler resets to prevent stale logits sampling (-2 error return).
- Expanded Behavioral Test Suite: Add fanged C++ and Java test assertions covering non-greedy distribution divergence across seeds, logits invalidation upon partial KV mutation, decoupled priming persistence, and struct alignment sizes.
- Bump project release version to v1.6.3 and update README documentation with state-machine invariant tables and memory layout offsets.

+++ bindings/java/libargus-core/src/main/java/cc/projectargus/libargus/ArgusSamplerConfig.java
@@ -16,7 +16,7 @@ package cc.projectargus.libargus;
  * @param dryBase            DRY exponential base (defaults to 1.75f)
  * @param dryAllowedLength   DRY allowed common n-gram length before penalty (defaults to 2)
  * @param dryPenaltyLastN    DRY penalty lookback window (-1 or 0 matches full context)
- * @param seed               RNG seed for distribution sampling (-1L matches random / LLAMA_DEFAULT_SEED)
+ * @param seed               RNG seed for distribution sampling (-1 matches random / LLAMA_DEFAULT_SEED)
  */
 public record ArgusSamplerConfig(
     float temperature,
@@ -31,7 +31,7 @@ public record ArgusSamplerConfig(
     float dryBase,
     int dryAllowedLength,
     int dryPenaltyLastN,
-    long seed
+    int seed
 ) {
 
     /**
@@ -45,7 +45,7 @@ public record ArgusSamplerConfig(
      * Creates a pure greedy / argmax sampling profile.
      */
     public static ArgusSamplerConfig greedy() {
-        return new Builder().temperature(0.0f).repeatPenalty(1.0f).topP(1.0f).minP(0.0f).topK(0).seed(-1L).build();
+        return new Builder().temperature(0.0f).repeatPenalty(1.0f).topP(1.0f).minP(0.0f).topK(0).seed(-1).build();
     }
 
     public static final class Builder {
@@ -61,7 +61,7 @@ public record ArgusSamplerConfig(
         private float dryBase = 1.75f;
         private int dryAllowedLength = 2;
         private int dryPenaltyLastN = -1;
-        private long seed = -1L;
+        private int seed = -1;
 
         public Builder() {}
 
@@ -123,7 +123,7 @@ public record ArgusSamplerConfig(
             return this;
         }
 
-        public Builder seed(long seed) {
+        public Builder seed(int seed) {
             this.seed = seed;
             return this;
         }
